requirements.txt
python-telegram-bot
openai
requests
pydantic